Bug#12661349 Assert in protocol::end_statement
The triggered assert checks that either OK or an error
is sent to the client after a command has been processed
by the server. It was triggered in this case if the client
used binary protocol and the connection was killed during
prepare. On release builds, the consequence of the bug was
that OK was sent to the client even if the prepare failed.

When processing a SQL command, a check towards end of
processing makes sure to report KILL_QUERY error to the
client if the connection/command was killed. This check
was however not done for binary protocol processing
(COM_STMT_PREPARE in this case).

This patch fixes the problem by making sure that the 
KILL_QUERY error is sent to the client in these cases
as well.

Test case added to ps.test.