BUG# 11757216: 49232: --STORAGE-ENGINE SHOULD BE DECOUPLED FROM DEFAULT TEMPTABLE TYPE
Implemented a --default-temp-storage_engine variable to complement the 
--default-storage-engine.
It's default value is the same as of --default-storage-engine.
But it's totally independent from --default-storage-engine.
The value of --default-temp-storage-engine affects only the tables created 
by CREATE TEMPORARY TABLE. All other internally created temp tables are not 
affected by it.
The special storage engine name "DEFAULT", when applied to such temporary
tables work as an alias of the default temp table storage engine.
Extended mysql-test-run.pl to issue --default-temp-storage-engine=myisam
as a complement to --default-storage-engine=myisam.

Test case added. 
Affected test cases updated.