Bug#12694147 COMPILING WITH -DWITHOUT_PARTITION_STORAGE_ENGINE=1              AND WITHOUT -O FAILS! (formerly known as 61625)
The problem was that the server would not compile if partitioning
was disabled. This was caused by two calls to the 
partition_key_modified() function which is only defined if
partitioning support is enabled. This regression was caused by
the patch for Bug#11762751.

This patch fixes the problem by wrapping the offending calls in
#ifdef WITH_PARTITION_STORAGE_ENGINE / #endif.

The patch also fixes a test case which used partitioning without
having "--source include/have_partition.inc". This meant that the
test case would fail on servers configured without partitioning.

No test case added.