WL#5915 - Make deadlock detection non-recursive.
The explicit stack is allocated at startup and freed at shutdown. The stack
size is set to OS_THREAD_MAX_N. If we reach this limit during the deadlock
search.  The boundary condition is flagged as a "search too deep" error and
the joining transaction that is requesting a lock in an incompatible mode is
rolled back.  This is similar to the "search too deep" behaviour of the
current algorithm. This patch also preserves the old behaviour of "search
too deep" where we also count the number of operations. The old behaviour
is now redundant and will be removed in a future version.
  
rb://617 Approved by Marko.