Bug#12603141: JOIN::flatten_subqueries asrt/simplify_joins sig11/... Bug#12603457: SEGFAULT IN REINIT_STMT_BEFORE_USE
These bugs are (almost) duplicates (depends empty/non-empty tables)

When preparing a prepared statement, we need to save a copy of
WHERE and HAVING condition trees in the prep_where and prep_having
fields of SELECT_LEX. However, the semantic analysis needs to wrap
some Item_field objects using Item_ref objects, which are scoped for
the lifetime of the preparation. This is usually not a problem,
but when the Item_field is the root object of the condition, the
Item_ref object is attempted saved in prep_where or prep_having.
The destructor for the Item_ref is called at the end of the
preparation, so that when execution of the statement is performed,
the destroyed Item_ref object will be referenced.