Bug #11757397 49437: CANNOT DO SHOW FIELDS FOR BIG VIEW
The problem was that views with too many columns in some
cases would appear to work correctly, while in other cases gave
various error messages. For example, DESCRIBE and SHOW commands
could give "Incorrect key file for table..." errors.

Depending on column datatypes used, views are materialized
as either MyISAM or Memory internal temporary tables.
The root cause of the problem was that MyISAM tables support
a limited number of columns. This is due too a 64K file header
size limitation. This limitation was not checked by CREATE VIEW.
Therefore views with unsupported sizes could be made, leading to
various errors when the view was later used.

This patch fixes the problem by reporting ER_TOO_MANY_FIELDS
error ("Too many columns") if views with more than 4096 columns
are created. Thus ensuring that any MyISAM table which might
be created for view processing will fit into the 64K header limit.
Note that 4096 was chosen as limit as this number is already the
limit on the number of columns for base tables.

Test case added to view.test.