WL#2111: #3 Implement GET DIAGNOSTICS command.
Introduce a set of classes to represent the skeleton of a GET
DIAGNOSTICS command.

The Sql_cmd_get_diagnostics class represents a GET DIAGNOSTICS
statement. The Diagnostics_information class represents diagnostics
information to be obtained through a GET DIAGNOSTICS command. The
Diagnostics_information_item class represents a single diagnostics
information item and is used to associate a specific diagnostics
information item to a target variable.

Within the execution context of the GET DIAGNOSTICS command,
implement the required special error handling. A GET DIAGNOSTICS
statement only fails if there are fatal errors, non-fatal errors
are simply appended to the message list.