WL#2111: #4 Implement GET DIAGNOSTICS statement information.
Implement the ability to retrieve statement information items
through a GET DIAGNOSTICS statement.

The Statement_information class is a specialization of
Diagnostics_information that aggregates statement information items,
which in turn are represented by the Statement_information_item class.

A statement information item is composed of a name and a member
function that obtains the value of the named item. Only the NUMBER
and ROW_COUNT statement information items are supported.