WL#2111: #5 Implement GET DIAGNOSTICS condition information.
Implement the ability to retrieve condition information items
through a GET DIAGNOSTICS statement.

The Condition_information class is a specialization of
Diagnostics_information that aggregates the condition for which
information is to be obtained and condition information items.
Items are represented by the Condition_information_item class.

A condition information item is composed of a name and a member
function that obtains the value of the named item. The supported
items match information that is currently available in a condition.