BUG#12604951 - RBR BREAKS IF SQL THREAD IS RESTARTED WHILE EXECUTING AN EVENT
This a regression caused by BUG#11763126 what introduced new flags and
procedures to track down changes that are unsafe to be rolled back.
In particular, after BUG#11763126 the trans_commit_stmt() propagates
unsafe rollback flags from statement to transaction level.

However, we cannot use this at the slave when row format is configured
as several events can be processed before trans_commit_stmt() is called.
This happens because trans_commit_stmt() is called only when the latest
event generated by a statement is processed.

Unfortunately, there are upper level functions that execute per event
and check transaction's status. So if the unsafe rollback flags are not
propagated at the end of every event, this can lead to errors.

To fix the problem, we propagate unsafe rollback flags from statement to
transaction level after processing every event if row format is in use.