Bug#12546960 - 60993: NAME QUOTED WITH QUOTE INSTEAD OF BACKTICK                GIVES NO SYNTAX ERROR
If a statement ended with mismatched quotes, the server would
accept the statement and present whatever was after the
initial quote as a text string.

In the particular bug report, the string was started with a
backtick.  Since the concluding quote was a regular 'tick',
it was not detected as and ending quote, so the tick was
added to the string.

The bug was not present when used interactively from the
mysql client, nor from mysqltest, as these tools do some
rudimentary quotes checking before sending the statement
to the server. 

The fix introduces a check to verify that the statement
does not end while still expecting an end quote.  If so - 
a parser error is returned.