Fix for Bug#12619744 MAIN.INNODB_MRR_NONE/INNODB_MRR_ICP FAILS OCCASIONALLY ON PB2
This is a test only problem. Due to InnoDB sometimes produces statistics for
number of rows in the table that are more than 50 percent off from the real
number this could change the execution strategy for the query to be
done as a table scan instead of a range scan.

To ensure that we get a stable plan for the query and that the query always
is executed as a range scan (which is what the test should test) this fix
adds a force index to the query in order to avoid that it can be executed
as a table scan.