Bug#12346211 - PERF IMPACT OF PERFORMANCE-SCHEMA WITH DISABLED INSTRUMENTS/CONSUMERS
Performance improvement #3: Better allocation.

Before this fix, the performance schema allocation algorithm used,
for example when creating a mutex, was:
- compute a random number I between 0 and N
- start to scan for an empty slot at position I modulo N,
 I+1 modulo N, I+2 modulo N ...
and this for each competing thread.

When only a few threads are running, and when the slots are mostly empty,
this works well since few collisions happen.

When the number of competing threads augments, and when the slots are mostly
occupied, this code can potentially create a lot of competition between
threads, when doing atomic operations against each slot internal lock.

With this fix, the allocation is changed to:
- keep a static monotonic counter for each buffer, I
- atomically increment I
- look for an empty slot at I modulo N
- repeat if necessary.

Because the monotonic counter is shared between all threads,
competing threads end up actually looking for different slots at I0, I0+1,
I0+2 etc, which prevents competition on each slot internal lock.
Also, because records are allocated in circular fashion, starting to look
for an empty slot after the last slot found increases the probability to
find available slots, and reduces the number of loops during the scan.

While difficult to measure, this fix is expected to improve the scalability
of the performance schema allocation code, which is used for:
- creating a mutex,
- creating a rwlock,
- creating a cond,
- opening a file,
- opening a table
which are all operations that can occur frequently when the server is
running.