Fixed bug#11765413 (formerly known as 58378): CRASH WITH DEPENDENT SUBQUERY AND PREPARED STATEMENT.
This crash occurred during execution of prepared statement that contained
subquery in WHERE clause and when this subquery used HAVING condition.

The problem was that during parsing of HAVING clause the instance of Item_field
was created on memory allocated from runtime memroot and this allocated
object/memory was destroyed after preparation stage finished. So when prepared
statement subsequently was executed it tried to use Item_field object that
had been already deleted.

The solution is to allocate Item_field object at memroot dedicated for prepared 
statement processing.