Bug#12795555: Missing rows with ANY subquery.
Item_maxmin_subselect::was_values was assigned to TRUE initially. It expected
to be reset on the first subselect execution, but when outer value is NULL
this doesn't happen, thus causing missing rows.
The solution is to initialize Item_maxmin_subselect::was_values to false and 
evaluate the subquery if hasn't been evaluated yet.

The patch is implemented by Evgeny Potemkin.