BUG#12798270 - ASSERTION `!TAB->SORTED' FAILED IN JOIN_READ_KEY2
An ASSERT that ordering is not requested on a table when eq_ref 
is used to access it was added by bug 11764737. The rationale 
is that eq_ref will only return a single row, so ordering 
becomes meaningless.

In this bug, the ASSERT triggered because the optimizer did 
request ordering on a table accessed through eq_ref. This 
request for ordered output has been there since long but 
surfaced due to the newly added ASSERT.

The fix is in make_join_readinfo() and is to not require an 
index to return rows in sorted order when eq_ref access 
method is chosen.