Bug 12831587 DOUBLE COPY IN NON-ICP CODE PATH IN ROW_SEARCH_FOR_MYSQL()
Due to the changes required by ICP we first copy a row from the InnoDB
format to the MySQL row buffer and then copy it to the pre-fetch queue.
This was done for the non-ICP code path too. This change removes the
double copy for the latter.

Note: We copy from the pre-fetch queue to the MySQL row buffer on exit
from the function row_search_for_mysql(). This is true for < trunk too.
If and when we fix ICP code to avoid double copy we should avoid the
copy from the InnoDB pre-fetch queue to the MySQL row buffer too. This will
require allocating memory for the pre-fetch queue to be done from
MySQL. That way we queue the row pointers instead of the contents.

rb://698 Approved by Jimmy Yang.