SPJ: Fixed 'of by one' bug in ::verifySortOrder() which caused the last element m_activeFrags[] to not be verified.
Also fixed ::fetchRow to used not advance the current row in
NdbReceiver which would cause the above verifySortOrder to fail.

NOTE: All NdbReceiver navigation in NdbQueryOperation use random
access to the received rows due to parent / child correlated 
results and therefore does not need ::get_row() which advance
to the next row.