SPJ: Refactoring of how & when the parent/child row correlation is handled:
Previously the correlation id's was decoded, and the parent/child 
hashMaps built as part of execTRANSID_AI().

This work is now deferred until a complete batch of result rows has been
received - and at this point it is handled by the thread which was waiting
for the result batch instead of the receiver thread.

Note: This patch is also a step in the direction of moving building
of the correlation hashmap entirely outside transporter mutex protected
code.