Bug #11746029: 22615: MYSQL SERVER INCORRECTLY CATOGORIZES THE LOST+FOUND DIRECTORY AS A DATABA
Implemented an --ignore-db-dir option that takes a valid file name or an empty string.
You can specify multiple options.
When a non-empty string is specified it's added as a directory name to the list of directories to ignore 
when executing SHOW commands and filling in INFORMATION_SCHEMA tables.
When an empty string is specified as an argument the list is reset to empty.
Note that the above list of ignored directories doesn't affect other SQL commands like e.g. USE or SELECT etc.
Added a read-only global @@ignore_db_dirs system variable that will contain the current list in effect.

In addition to the list above the server will always ignore the directories starting with a dot (.). This is safe 
because since mysql-5.1 a name of a database directory can't start with a dot (.) thanks to the fact that we
encode database and table names.

Added a test case.
Fixed the affected test cases.
Fixed the memory handling when converting the directories array to a hash fails for some reason.