WL#5758: Improve Thread Scheduling
Use atomics where available to enforce concurrency limit. The atomics
version of the scheduling code doesn't use an explicit queue that
guarantees eventual progress and a total order on the waiters. Instead
it uses os_thread_sleep() to force the threads to wait. ie. the newi
scheduling is not starvation free and imposes no order on the entering
transactions.
				          
Introduce new config variable:
					        
  innodb_adaptive_max_sleep_delay := 0..1000000 in micro-seconds

If the value is greater than 0 (which implies that it is enabled)  then
the innodb_thread_sleep_delay will be adjusted dynamically based on the
system load. The aim is to make it converge on the optimal sleep delay
when concurrency thread limit is active.

The new configuration variable is an upper limit that will be imposed
when determining the optimal value of innodb_thread_sleep_delay.

As in the old version which is enabled on platforms that don't support
atomic builtins this scheduling of threads will be activated based on
the settings of innodb_thread_concurrency.

The optimal value for innodb_thread_concurrency is difficult to standardise.
Experimental observations suggest that on hardware that has greater than
8 cores the setting it to the number of cores seems to work best. But on
hosts that have 8 cores (but with faster CPUs) the optimal value needs to
be determined empirically. On the hosts tested the best Sysbench results
were obtained with a setting innodb_thread_concurrency=96 for both RO/RW.

Aproved by Inaam/Marko rb://651