This commit implements adaptive parallelism for (non-root) scan operations (i.e. index scans). So far, such scans have always been executed with maximal parallelism, such that each fragment were scanned in parallel. This also meant that the available batch size would have to be divided by the number of fragments. For index scans with large result sets, this was inefficient, since it meant asking for many small batches.
As an example, assume that a table scan of t1 is followed by an index scan on 
t2 where the index has poor selectivity. The query is then effectively a cross
join. When such queries where tested on distributed clusters with 
multi-threaded ndbd, the queries would typically run five to six times slower 
than with query pushdown disabled.

This commit will therefore try to set an optimal parallelism, depending on the
size of the scan result set. This eliminates the performance regression in 
the test cases described above.

This works as follows:

* The first time an index scan is started, it starts with parallelism=1 for 
the first batch.

* For the subsequent batches, the SPJ block will try to set the parallelism 
such that all rows from a single fragment will fit in one batch. If this
is not possible, parallelism will remain 1.

* Whenever the index scan is started again (typically after receiving a new
batch for its parent), the SPJ block will try to guess the optimal parallelism
based on statistics from previous runs. 'Optimal' means that one can expect
all rows from a given fragment to be retrieved in one batch if possible.

See also quilt sereies published on 2011-08-18.