Fixed SPJ regression:
After push of 'revno: 4448' to branch 'mysql-5.1-telco-7.0', the ndbapi test 
'testSpj' started failing.

This is a fix for the first of these problems ('testSpj -n scanJoin' still fails)
which reverts part of the changes in revno 4448.

The problem fixed - or rather compensated for - is that retrieving ATTR values
from the receive buffer has the underlying assumption that the NdbReceiver
refer the next row (or end of the current).

Therefor this fix increment the 'current' row by using ::get_row instead of ::peek_row
and remove a verifySortOrder from a place where the current row has been advanced, and
before the fragment has been reorganized().