Fix for SPJ regression introduced by revno4456 to branch 'mysql-5.1-telco-7.0'
My previous 'correlation.patch' stores a TupleCorrelation (Uint32) together
with each received row directly in the NdbReceiver buffer.

This causes problem when there are RecAttr's in the received data.
NdbReceiver has an assumption that everything not being a NdbRecord
are NdbRecAttr in this buffer. It therefore try to parse this TupleCorrelation
as a NdbRecAttr which will fail.

Instead of storing the TupleCorrelations together with the row, we now
allocate a separate array of TupleCorrelations as part of each NdbResultSet
where the TupleCorrelations are stored.