WL#5986: Stored programs: make SQL-handlers closer to the standard).
Patch for Bug#12731619: NESTED SP HANDLERS CAN TRIGGER ASSERTION.
  
The thing is that runtime (execution) call stack can be deeper than the one
during the parsing stage. I.e. runtime call stack can be deeper than number of
nested BEGIN ... END blocks.

Before the patch, the code was written in an assumption, that depth of runtime
call stack can be calculated at parse time. That assumption led to a crash (when
it's not true).

The user visible problem was that the server might crash when
executing a stored program containing nested SQL-handlers.
During parsing, the maximum nesting was determined and
used to allocate memory for handler invocations. However, in
some cases this maximum depth could not be determined correctly.
This could cause the server to crash when the routine was executed.

The following things have been done in the patch:
  - Introduce Sql_condition_info to store 4 needed attributes
    from MYSQL_ERROR to report raised_condition;
  - Restructure find_handler() logic;
  - Get rid of static arrays for sp_rcontext::m_handlers and
    sp_rcontext::m_raised_condition;
  - Get rid of static handler-call-stack in sp_rcontext.
    Use dynamic call-stack.
  - Move deinitialization from sp_pcontext::destroy() to
    sp_pcontext destructor;
  - Remove sp_pcontext::destroy();
  - delete sp_pcontext instance in sp_head destructor;
  - Get rid of two-steps sp_rcontext initialization.
    Introduce static sp_rcontext::create() as a single point
    to construct a valid sp_rcontext instance.

sp_rcontext cleanup / polishing:
  - use Bounds_checked_array template to check access to "raw" arrays.