Fixed bug#11749050 (formerly known as 38028): Assert hit in function open_table_def in file table.cc
This bug relates only to the case when storage handlerton exposes
the flag HTON_FLUSH_AFTER_RENAME and user issues ALTER statement
that contains RENAME clause. In this case server, built in debug mode,
has crashed on assert.

The reason for this bug was that, when temporary table was opened before
flushing storage engine log, the old name of table was used. Since at
this point table with this name no longer existed assert was triggered.

The solution is to open new (renamed version of table) table before
flushing storage engine log.

Testcase for this bug is missing because none of currently supported
storage engines exposes the flag HTON_FLUSH_AFTER_RENAME.