Fix for Bug#12822678: 2 MORE ROWS WHEN ICP=ON W/ STRAIGHT_JOIN
The cause for the wrong result was that we pushed down index conditions
for a table with join type CONST. This table was the second table in
the join order. The pushed index condition contained a field from the
first table. Due to optimization of "const" tables the single row from
these are only read once from the storage engine and then re-used
multiple times during the join. This caused the pushed index condition
to only be evaluated once for the first access instead of on all
accesses to the row. This causes the pushed index condition to not
be evaluated for all records and can result in either more or zero
rows from the join.

The fix for this is to not use ICP when the join type is CONST
or SYSTEM. An alternative fix that also would solve this problem
is to only allow index conditions that only accessed fields from
the current table during evaluation. This would avoid index
conditions that referred to other tables to be pushed down.