This commit concerns pushed queries (i.e. using NdbQueryOperation etc.) with child scan operations.
This commit does the following:

1) Earlier, the api would try to set batch size (in bytes and rows) high enough 
to run child scans with full parallelism, even if this meant exceeding 
configuration limits (as set by BatchByteSize, BatchSize, and 
MaxScanBatchSize). The downside of this was that you could potentially get 
very large batches. (With 64 fragments and 10k records you could get 
64x64x10k=40M.)

Since sequential child scans (+ adaptive parallelism) have now been
implemented we no longer need to run child scans with full parallelism.
Therefore, we no longer need to exceed the limits from the configuration.
This commit  therefore use the batch size (in bytes and rows) that
NdbReceiver::calculate_batch_size() suggests. Child scans may then run with
reduced parallelism if the allowed number of rows is lower that the number
of fragments. (But the user may increase BatchByteSize and MaxScanBatchSize
to counter this.)

2) This commit changes how QueryNodeParameters::batchSize is used. Instead of
16 bits for number of rows and 16 for number of bytes, there is now 11 bits
for number of rows and 21 for number of bytes. The maximal values for the
BatchByteSize and BatchSize configuration variables are 1 megabyte and 992
rows, so 16 bits was not enough for representing all legal values (large values
were silently truncated).