Bug 12866669 - COMPILER HINTS CAUSE PERFORMANCE TO DROP
Make the compiler hints optional during a build. Introduce config variable:

  --innodb-compiler-hints := on|off

The default value is still ON. Changing it to OFF we will do separately. With
this change at least we have a mechanism to test whether they are useful or
not. So far they have been shown to cause minor regression instead of improving
performance. It is possible that they were useful in the past.

  rb://742 Approved by Jimmy Yang.